Backend_python/requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-groq
vconsoleprint


python-dotenv
pydantic

fastapi
uvicorn

requests
